aws-infra/requirements.txt
# For AWS Lambda function
boto3>=1.20

# For AWS CDK (infrastructure code)
aws-cdk-lib>=2.0.0
constructs>=10.0.0